Add a gallery for the iOS NavigationPage stuff
