+++ Xamarin.Forms.Controls.Issues/Xamarin.Forms.Controls.Issues.Shared/PlatformSpecifics_iOSTranslucentNavBarX.xaml.cs
@@ -25,7 +25,7 @@ namespace Xamarin.Forms.Controls.Issues
 		{
 			var button = new Button { Text = "Toggle Translucent", BackgroundColor = Color.Yellow };
 
-			button.Clicked += (sender, args) => On<iOS>().EnableTranslucentNavigationBar(!On<iOS>().IsNavigationBarTranslucent());
+			button.Clicked += (sender, args) => On<iOS>().SetIsNavigationBarTranslucent(!On<iOS>().IsNavigationBarTranslucent());
 
 			var content = new ContentPage
 			{
